Rename: /api/openapi -> api/docs endpoint
