+++ server/api/service/docs/controller.test.js
@@ -3,9 +3,9 @@ import request from 'supertest';
 import SwaggerParser from 'swagger-parser';
 import server from '../../../server';
 
-describe('api/openapi', () => {
+describe('api/docs', () => {
   test('responds as expected', async () => {
-    const res = await request(server).get('/api/openapi');
+    const res = await request(server).get('/api/docs');
     expect(res.status).toEqual(200);
     expect(res.type).toMatch(/json/);
     expect(res.body).toMatchObject({
